# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FunLand MC API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
  isAdmin: boolean;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  codeVerifier: string;
  state: string;
  nonce: string;
  redirectUri: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ServerStatusStatus =
  (typeof ServerStatusStatus)[keyof typeof ServerStatusStatus];

export const ServerStatusStatus = {
  online: "online",
  offline: "offline",
  maintenance: "maintenance",
} as const;

export interface ServerStatus {
  status: ServerStatusStatus;
  playersOnline: number;
  maxPlayers: number;
  ip: string;
  port: number;
  version: string;
  message?: string | null;
  updatedAt: string;
}

export type UpdateServerStatusBodyStatus =
  (typeof UpdateServerStatusBodyStatus)[keyof typeof UpdateServerStatusBodyStatus];

export const UpdateServerStatusBodyStatus = {
  online: "online",
  offline: "offline",
  maintenance: "maintenance",
} as const;

export interface UpdateServerStatusBody {
  status?: UpdateServerStatusBodyStatus;
  playersOnline?: number;
  maxPlayers?: number;
  ip?: string;
  port?: number;
  version?: string;
  message?: string | null;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  coins: "coins",
  rank: "rank",
  unban: "unban",
} as const;

export interface Product {
  id: string;
  category: ProductCategory;
  name: string;
  priceInr: number;
  tagline?: string | null;
  coins?: number | null;
  rankTier?: string | null;
  unbanDuration?: string | null;
  accent?: string | null;
  sortOrder: number;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  discord: "discord",
  website: "website",
} as const;

export interface CreateOrderBody {
  productId: string;
  minecraftUsername: string;
  contact?: string | null;
  referral?: string | null;
  paymentMethod: CreateOrderBodyPaymentMethod;
  paymentProofPath?: string | null;
  transactionRef?: string | null;
  notes?: string | null;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  discord: "discord",
  website: "website",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  verified: "verified",
  delivered: "delivered",
  rejected: "rejected",
} as const;

export interface Order {
  id: string;
  productId: string;
  productName: string;
  productCategory: string;
  priceInr: number;
  minecraftUsername: string;
  contact?: string | null;
  referral?: string | null;
  paymentMethod: OrderPaymentMethod;
  paymentProofPath?: string | null;
  transactionRef?: string | null;
  notes?: string | null;
  status: OrderStatus;
  adminNote?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type UpdateAdminOrderBodyStatus =
  (typeof UpdateAdminOrderBodyStatus)[keyof typeof UpdateAdminOrderBodyStatus];

export const UpdateAdminOrderBodyStatus = {
  pending: "pending",
  verified: "verified",
  delivered: "delivered",
  rejected: "rejected",
} as const;

export interface UpdateAdminOrderBody {
  status: UpdateAdminOrderBodyStatus;
  adminNote?: string | null;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  deliveredOrders: number;
  revenueInr: number;
  recentOrders: Order[];
}

export type ExchangeMobileAuthorizationCode200 = {
  sessionToken: string;
  user: AuthUser;
};