# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FunLand MC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionToken: { type: string }
                  user: { $ref: "#/components/schemas/AuthUser" }
                required: [sessionToken, user]
...
  /admin/orders/{orderId}:
    patch:
      operationId: updateAdminOrder
      tags: [admin]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string, nullable: true }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        profileImageUrl: { type: string, nullable: true }
        isAdmin: { type: boolean }
      required: [id, isAdmin]
...
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code: { type: string }
        codeVerifier: { type: string }
        state: { type: string }
        nonce: { type: string }
        redirectUri: { type: string }
      required: [code, codeVerifier, state, nonce, redirectUri]
...
    Order:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        productCategory: { type: string }
        priceInr: { type: integer }
        minecraftUsername: { type: string }
        contact: { type: string, nullable: true }
        referral: { type: string, nullable: true }
        paymentMethod: { type: string, enum: [discord, website] }
        paymentProofPath: { type: string, nullable: true }
        transactionRef: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        status: { type: string, enum: [pending, verified, delivered, rejected] }
        adminNote: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, productId, productName, productCategory, priceInr, minecraftUsername, paymentMethod, status, createdAt, updatedAt]
    UpdateAdminOrderBody:
      type: object
      properties:
        status: { type: string, enum: [pending, verified, delivered, rejected] }
        adminNote: { type: string, nullable: true }
      required: [status]
...