# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FunLand MC API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
      isAdmin: zod.boolean(),
    }),
    zod.null(),
  ]),
});

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string(),
  codeVerifier: zod.string(),
  state: zod.string(),
  nonce: zod.string(),
  redirectUri: zod.string(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  sessionToken: zod.string(),
  user: zod.object({
    id: zod.string(),
    email: zod.string().nullish(),
    firstName: zod.string().nullish(),
    lastName: zod.string().nullish(),
    profileImageUrl: zod.string().nullish(),
    isAdmin: zod.boolean(),
  }),
});

export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});

export const GetServerStatusResponse = zod.object({
  status: zod.enum(["online", "offline", "maintenance"]),
  playersOnline: zod.number(),
  maxPlayers: zod.number(),
  ip: zod.string(),
  port: zod.number(),
  version: zod.string(),
  message: zod.string().nullish(),
  updatedAt: zod.coerce.date(),
});

export const UpdateServerStatusBody = zod.object({
  status: zod.enum(["online", "offline", "maintenance"]).optional(),
  playersOnline: zod.number().optional(),
  maxPlayers: zod.number().optional(),
  ip: zod.string().optional(),
  port: zod.number().optional(),
  version: zod.string().optional(),
  message: zod.string().nullish(),
});

export const UpdateServerStatusResponse = zod.object({
  status: zod.enum(["online", "offline", "maintenance"]),
  playersOnline: zod.number(),
  maxPlayers: zod.number(),
  ip: zod.string(),
  port: zod.number(),
  version: zod.string(),
  message: zod.string().nullish(),
  updatedAt: zod.coerce.date(),
});

export const ListProductsResponseItem = zod.object({
  id: zod.string(),
  category: zod.enum(["coins", "rank", "unban"]),
  name: zod.string(),
  priceInr: zod.number(),
  tagline: zod.string().nullish(),
  coins: zod.number().nullish(),
  rankTier: zod.string().nullish(),
  unbanDuration: zod.string().nullish(),
  accent: zod.string().nullish(),
  sortOrder: zod.number(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

export const CreateOrderBody = zod.object({
  productId: zod.string(),
  minecraftUsername: zod.string(),
  contact: zod.string().nullish(),
  referral: zod.string().nullish(),
  paymentMethod: zod.enum(["discord", "website"]),
  paymentProofPath: zod.string().nullish(),
  transactionRef: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const CreateOrderResponse = zod.object({
  id: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  productCategory: zod.string(),
  priceInr: zod.number(),
  minecraftUsername: zod.string(),
  contact: zod.string().nullish(),
  referral: zod.string().nullish(),
  paymentMethod: zod.enum(["discord", "website"]),
  paymentProofPath: zod.string().nullish(),
  transactionRef: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.enum(["pending", "verified", "delivered", "rejected"]),
  adminNote: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

export const ListAdminOrdersResponseItem = zod.object({
  id: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  productCategory: zod.string(),
  priceInr: zod.number(),
  minecraftUsername: zod.string(),
  contact: zod.string().nullish(),
  referral: zod.string().nullish(),
  paymentMethod: zod.enum(["discord", "website"]),
  paymentProofPath: zod.string().nullish(),
  transactionRef: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.enum(["pending", "verified", "delivered", "rejected"]),
  adminNote: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListAdminOrdersResponse = zod.array(ListAdminOrdersResponseItem);

export const UpdateAdminOrderParams = zod.object({
  orderId: zod.coerce.string(),
});

export const UpdateAdminOrderBody = zod.object({
  status: zod.enum(["pending", "verified", "delivered", "rejected"]),
  adminNote: zod.string().nullish(),
});

export const UpdateAdminOrderResponse = zod.object({
  id: zod.string(),
  productId: zod.string(),
  productName: zod.string(),
  productCategory: zod.string(),
  priceInr: zod.number(),
  minecraftUsername: zod.string(),
  contact: zod.string().nullish(),
  referral: zod.string().nullish(),
  paymentMethod: zod.enum(["discord", "website"]),
  paymentProofPath: zod.string().nullish(),
  transactionRef: zod.string().nullish(),
  notes: zod.string().nullish(),
  status: zod.enum(["pending", "verified", "delivered", "rejected"]),
  adminNote: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

export const GetAdminStatsResponse = zod.object({
  totalOrders: zod.number(),
  pendingOrders: zod.number(),
  deliveredOrders: zod.number(),
  revenueInr: zod.number(),
  recentOrders: zod.array(
    zod.object({
      id: zod.string(),
      productId: zod.string(),
      productName: zod.string(),
      productCategory: zod.string(),
      priceInr: zod.number(),
      minecraftUsername: zod.string(),
      contact: zod.string().nullish(),
      referral: zod.string().nullish(),
      paymentMethod: zod.enum(["discord", "website"]),
      paymentProofPath: zod.string().nullish(),
      transactionRef: zod.string().nullish(),
      notes: zod.string().nullish(),
      status: zod.enum(["pending", "verified", "delivered", "rejected"]),
      adminNote: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});