# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FunLand MC API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  productId: string;
  productName: string;
  productCategory: string;
  priceInr: number;
  minecraftUsername: string;
  contact?: string | null;
  referral?: string | null;
  paymentMethod: OrderPaymentMethod;
  paymentProofPath?: string | null;
  transactionRef?: string | null;
  notes?: string | null;
  status: OrderStatus;
  adminNote?: string | null;
  createdAt: Date;
  updatedAt: Date;
}