# lib/api-zod/src/generated/types/updateAdminOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FunLand MC API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAdminOrderBodyStatus } from "./updateAdminOrderBodyStatus";

export interface UpdateAdminOrderBody {
  status: UpdateAdminOrderBodyStatus;
  adminNote?: string | null;
}